requirements.txt
streamlit
python-dotenv
google-generativeai
requests
pdfplumber
python-docx
newspaper3k
transformers
torch
serpapi

